Refactor People Wrapper
