+++ src/index.js
@@ -2,8 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {
     BrowserRouter as Router,
-    Switch,
-    Route,
 } from "react-router-dom";
 import './index.css';
 import {applyMiddleware, createStore} from 'redux'
